* v1.37d, 20 April 2006       + Added "Enlarge Image to Fit Result" option to Image/Rotate/         Arbitrarily.       + Added "Case Sensitive" and "Whole Words" options to the text         editor's Edit/Find command.       + Added "Rodbard (NIH Image)" to the list of functions in Analyze/         Calibrate.       + The flood fill (paint bucket) macro tool is now recorded.       + Global variables can now be defined anywhere in a macro file         outside of macro and function blocks.       + The drawString() macro function uses antialiasing.       + Macros with names ending in "Tool Selected" are not installed in         the menu bar.       + Added constructors to the ImagePlus and ShortProcessor classes that         support 16-bit BufferedImages.       + Fixed a bug that caused File/Save As/Image Sequence to not transfer         calibration and DICOM tags to files saved in TIFF format.       + Fixed a bug that sometimes caused incorrect histograms of         calibrated 8-bit images and signed 16-bit images.       + Fixed a bug that caused incorrect histograms of images containing         line, point or angle selections.       + Fixed a bug that caused histograms of images with more than 16         million pixels with the same value to be drawn incorrectly.       + Fixed a bug that sometimes caused Image/Stacks/Z Project to produce         incorrect "Standard Deviation" output with 16 and 32 bit images.       + Fixed bug that caused text to be drawn incorrectly on contrast         enhanced 8-bit images.       + Fixed a bug in File/Import/Image Sequence that caused a bogus "ZVI         Reader plugin not found" error message when opening files from a         directory containing a "Thumbs.db" file.       + Fixed a bug that caused opening of images from the command line to         fail after use of the "cd" command.       + Fixed a bug, introduced in 1.37b, that caused null characters in         Image/Show Info output of DICOM images.

+++ ij/gui/HistogramWindow.java
@@ -213,7 +213,7 @@ public class HistogramWindow extends ImageWindow implements Measurements, Action
 		int index, y;
 		for (int i = 0; i<HIST_WIDTH; i++) {
 			index = (int)(i*(double)histogram.length/HIST_WIDTH); 
-			y = (int)(HIST_HEIGHT*histogram[index])/maxCount;
+			y = (int)((double)HIST_HEIGHT*histogram[index])/maxCount;
 			if (y>HIST_HEIGHT)
 				y = HIST_HEIGHT;
 			ip.drawLine(i+XMARGIN, YMARGIN+HIST_HEIGHT, i+XMARGIN, YMARGIN+HIST_HEIGHT-y);
